requirements.txt
Pillow
PyExifTool
